S10P22A303-353 [feat] 키바나 대시보드 구성 위한 로그 구성 * S10P22A303-353 [feat] 키바나 대시보드 구성 위한 로그 구성

+++ backend/MissionMS/src/main/java/com/a303/missionms/domain/dailyMission/service/DailyMissionServiceImpl.java
@@ -200,7 +200,7 @@ public class DailyMissionServiceImpl implements DailyMissionService {
 		List<MissionLogRes> missionLogList = dailyMissionRepository.findAllToMissionLogRes();
 
 		// 미션 수행도 집계
-		int percent = 0;
+		float percent = 0;
 		LocalDate localDate = LocalDate.now().minusDays(1);
 		for (MissionLogRes missionLogRes : missionLogList) {
 			if (missionLogRes.isFinish()) {
